create module PrepareNetwork with source creation from hydrologic network
